+++ plugins/obsidian/evals/run.py
@@ -24,6 +24,7 @@ EVAL_MODULES = (
     "eval_distill_placement",
     "eval_retrieval_verification_report",
     "eval_dlq_on_missing_scores",
+    "eval_search_parity",
 )